Corrects to how restarts are handled.
More fixes still need to be made; I'm attempting to pass around _only_ a
reference to the data in the HDF5 file until it is absolutely essential
that we have the data in memory.  Doing so means there is currently some
rather careful handling that must be done.

I'm sure we can make adjustments to make this more appropriate and
faster, however.